refactor : move header into widget
